src/services/solana-service.ts
import { Connection, PublicKey, TransactionInstruction, Transaction, SystemProgram } from '@solana/web3.js';

class SolanaService {
  private connection: Connection;
  private walletAddress: PublicKey;
  private authorityPrivateKey: Uint8Array;
  private programId: PublicKey; // Add the program's public key
  private tokenAccountAddress: PublicKey; // Add the token account's public key

  constructor(connectionUrl: string, walletAddress: string, authorityPrivateKey: Uint8Array) {
    this.connection = new Connection(connectionUrl, 'singleGossip');
    this.walletAddress = new PublicKey(walletAddress);
    this.authorityPrivateKey = authorityPrivateKey;
    this.programId = new PublicKey('TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA'); // Program's public key
    this.tokenAccountAddress = new PublicKey('5P3giWpPBrVKL8QP8roKM7NsLdi3ie1Nc2b5r9mGtvwb'); // Token account's public key
  }

  async fetchData(): Promise<any> {
    // Implement fetching data from the TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA program and token account
    // Example: Use Solana's Token API to interact with the program and account
    // Return the data you retrieve
  }

  async sendTransaction(instructions: TransactionInstruction[]): Promise<string> {
    // Create a transaction and add instructions
    const transaction = new Transaction();
    instructions.forEach((instruction) => transaction.add(instruction));

    // Sign the transaction
    transaction.feePayer = this.walletAddress;
    transaction.recentBlockhash = (await this.connection.getRecentBlockhash()).blockhash;
    transaction.sign(this.authorityPrivateKey);

    // Send the transaction
    const transactionHash = await this.connection.sendTransaction(transaction);

    return transactionHash;
  }
}

export default SolanaService;
