src/tests/solana-service.test.ts
import { Connection, TransactionInstruction, Transaction, PublicKey } from '@solana/web3.js';
import SolanaService from './solana-service'; // Import your Solana service here

// Mocked Solana connection for testing
const mockedConnection = new Connection('https://api.devnet.solana.com', 'singleGossip');

// Mocked wallet private key for testing
const mockedPrivateKey = new Uint8Array([/* your byte values here */]);

// Initialize your Solana service with the mocked connection
const solanaService = new SolanaService('https://api.devnet.solana.com', '5P3giWpPBrVKL8QP8roKM7NsLdi3ie1Nc2b5r9mGtvwb', mockedPrivateKey);

describe('Solana Service Tests', () => {
  beforeAll(() => {
    // Setup any necessary configurations or test data
  });

  afterAll(() => {
    // Cleanup resources if needed
  });

  test('Fetch Data from Solana', async () => {
    const data = await solanaService.fetchData();

    // Write your assertions to test the fetched data
    expect(data).toBeDefined();
    // Add more assertions based on your specific data structure and expectations
  });

  test('Send Transaction to Solana', async () => {
    const instructions: TransactionInstruction[] = [
      // Define your transaction instructions here
      // Example: Create, update, or interact with a Solana program
    ];

    const transactionHash = await solanaService.sendTransaction(instructions);

    // Write your assertions to test the transaction result
    expect(transactionHash).toBeDefined();
    // Add more assertions based on your specific transaction expectations
  });
});
